+++ src/generic-table/interfaces/gt-config-setting.ts
@@ -5,5 +5,6 @@ export interface GtConfigSetting {
   sort?:string,
   sortOrder?:number,
   columnOrder?:number,
-  export?:boolean // should column be exported to CSV (OPTIONAL)
+  export?:boolean, // should column be exported to CSV (OPTIONAL)
+  search?:boolean // should we include this column when using global search (OPTIONAL)
 }